requirements.txt
absl-py==1.1.0
altair==4.2.0
appnope==0.1.2
asgiref==3.4.1
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.9.3
black==21.12b0
blinker==1.5
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.3
click==7.1.2
cloudbiolinux==0.1
cmake==3.18.4.post1
commonmark==0.9.1
contrib==0.3.0
cryptography==37.0.2
cycler==0.10.0
decorator==5.1.1
Django==3.2.5
dlib==19.18.0
entrypoints==0.4
executing==0.8.2
fabric==2.7.0
face-recognition==1.3.0
face-recognition-models==0.3.0
face-recognition-util==0.1.2
Flask==1.1.2
flatbuffers==22.12.6
gast==0.4.0
gevent==21.1.2
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
google-search-results==2.3.0
google-search-results-serpwow==1.1.11
greenlet==1.0.0
grpcio==1.34.1
h5py==3.1.0
idna==3.2
imageio==2.19.3
importlib-metadata==4.10.0
invoke==1.7.1
ipython==8.0.0
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.3
joblib==1.1.0
jsonschema==4.6.0
keras==2.11.0
Keras-Applications==1.0.4
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.0.2
kiwisolver==1.3.1
libclang==14.0.1
lxml==4.6.3
Markdown==3.3.6
MarkupSafe==1.1.1
matplotlib==3.3.4
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
networkx==2.8.4
numpy==1.22.4
oauthlib==3.1.1
opencv-python==4.6.0.66
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
paramiko==2.11.0
parso==0.8.3
pathlib2==2.3.7.post1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.4.1
prompt-toolkit==3.0.24
protobuf==3.19.3
ptyprocess==0.7.0
pure-eval==0.2.1
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
Pygments==2.11.2
Pympler==1.0.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.1
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rich==12.6.0
rsa==4.8
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
soupsieve==2.2.1
sqlparse==0.4.1
stack-data==0.1.3
streamlit==1.15.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-aarch64==1.2
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2022.5.4
toml==0.10.2
tomli==1.2.3
toolz==0.12.0
torch==1.10.2
torchvision==0.11.3
tornado==6.2
traitlets==5.1.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.6
validators==0.20.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.2.0
